Backend/Mask_RCNN-master/requirements.txt
absl-py==1.4.0
astor==0.8.1
backcall==0.2.0
cached-property==1.5.2
click==8.0.4
colorama==0.4.5
cycler==0.11.0
Cython==3.0.9
dataclasses==0.8
decorator==4.4.2
entrypoints==0.4
Flask==2.0.3
Flask-Cors==4.0.0
gast==0.5.4
grpcio==1.48.2
h5py==2.10.0
imageio==2.15.0
imgaug==0.4.0
importlib-metadata==4.8.3
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.3
jupyter-client==7.1.2
jupyter-core==4.9.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
MarkupSafe==2.0.1
mask-rcnn==2.1
matplotlib==3.3.4
mock==5.1.0
nest-asyncio==1.6.0
networkx==2.5.1
numpy==1.19.5
opencv-python==4.9.0.80
parso==0.7.1
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.36
protobuf==3.19.6
Pygments==2.14.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyWavelets==1.1.1
pywin32==305
PyYAML==6.0.1
pyzmq==25.1.2
scikit-image==0.17.2
scipy==1.5.4
Shapely==1.8.5.post1
six==1.16.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
tifffile==2020.9.3
tornado==6.1
traitlets==4.3.3
typing_extensions==4.1.1
wcwidth==0.2.13
websockets==9.1
Werkzeug==2.0.3
zipp==3.6.0
